Use AtomicUsize for read limiter + feature gate it
