+++ capnp/src/private/mod.rs
@@ -29,6 +29,7 @@ mod primitive;
 pub mod layout;
 mod mask;
 pub mod units;
+mod read_limiter;
 mod zero;
 
 #[cfg(test)]
